ml_service/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
numpy
Pillow
ultralytics
python-multipart
